pyproject.toml
[project]
name = "src"
version = "0.0.1"
description = ""

readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "langgraph>=1.0.0",
    "python-dotenv>=1.0.1",
    "langchain (>=1.2.7,<2.0.0)",
    "langchain-text-splitters (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-openai (>=1.1.7,<2.0.0)",
    "langchain-core (>=1.2.7,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "langchain-astradb (>=1.0.0,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "langchain-google-genai (>=4.2.0,<5.0.0)",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"




[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.4.7",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
